backend/requirements.txt
discord.py>=2.3.2
ffmpeg-python
pydub
requests
yt-dlp
Pillow
aiohttp
python-dotenv
psutil
PyNaCl
gtts
flask
gunicorn
spotipy
psycopg2-binary
librosa
scipy
numpy
soundfile
stripe
numpy
flask_cors